pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pdf-converter"
authors = [{name = "Clément Dubos", email = "contact@cdubos.fr"}]
dynamic = ["version", "description"]
readme = "README.md"
requires-python = ">= 3.9"
dependencies = ['typer[all]', 'pillow', 'pdf2image']

[project.scripts]
pdf-converter = "pdf_converter:cli"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
    "pdf_converter.py",
    "pyproject.toml:version"
]
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]
